# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/analytics/dashboard": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Dashboard analytics (khusus admin)",
                "description": "Statistik umum untuk dashboard admin",
                "operationId": "60d2bed1d73ca56b203b6e5ed729ca67",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_users": {
                                            "type": "integer"
                                        },
                                        "total_berita": {
                                            "type": "integer"
                                        },
                                        "total_likes": {
                                            "type": "integer"
                                        },
                                        "total_views": {
                                            "type": "integer"
                                        },
                                        "users_by_role": {
                                            "type": "object"
                                        },
                                        "users_by_membership": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/analytics/users": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Analytics user (khusus admin)",
                "description": "Statistik detail tentang user",
                "operationId": "7f4e4f64239eda36ff5d468197d2ca93",
                "responses": {
                    "200": {
                        "description": "User analytics"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/analytics/content": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Analytics konten (khusus admin)",
                "description": "Statistik detail tentang konten berita",
                "operationId": "044d7ecde74caa2ad910a585e200676e",
                "responses": {
                    "200": {
                        "description": "Content analytics"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}